+++ Amor/View/View/DMView.swift
@@ -89,53 +89,6 @@ final class DMView: BaseView {
         spaceImageView.backgroundColor = .gray
     }
     
-    private func setDmCollectionViewLayout(_ type: DMCollectionViewType) -> UICollectionViewLayout {
-        
-        let layout: UICollectionViewLayout
-        
-        switch type {
-        case .user:
-            layout = setDMUserCollectionViewLayout()
-        case .chat:
-            layout = setDMChatCollectionViewLayout()
-        }
-        
-        return layout
-    }
-    
-    private func setDMUserCollectionViewLayout() -> UICollectionViewLayout {
-        let itemSize = NSCollectionLayoutSize(
-            widthDimension: .fractionalWidth(1),
-            heightDimension: .fractionalHeight(1)
-        )
-        
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-        
-        let groupSize = NSCollectionLayoutSize(
-            widthDimension: .absolute(80),
-            heightDimension: .fractionalHeight(1)
-        )
-        
-        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
-        
-        let section = NSCollectionLayoutSection(group: group)
-        section.orthogonalScrollingBehavior = .continuous
-        section.contentInsets = .init(top: 0, leading: 5, bottom: 0, trailing: 5)
-        
-        let layout = UICollectionViewCompositionalLayout(section: section)
-        
-        return layout
-    }
-    
-    private func setDMChatCollectionViewLayout() -> UICollectionViewLayout {
-        var listConfiguration = UICollectionLayoutListConfiguration(appearance: .plain)
-        listConfiguration.showsSeparators = false
-        
-        let layout = UICollectionViewCompositionalLayout.list(using: listConfiguration)
-        
-        return layout
-    }
-    
     override func layoutSubviews() {
         super.layoutSubviews()
